Added A0 management to GUI and derivedEnergy driver.  This involves multiple steps.  An OSA Focus scan focuses on the OSA and set cursor to focus then changes the ZonePlateZ offset so the focus position is at 0.  Then a sample focus scan will leave the ZonePlateZ offset unchanged but change the SampleZ offset such that it's position is equal to A0.  We don't do this at cosmic so there is a "calibrated A0" flag in the config that turns these features on and off.
Added a new version of the GUI which is only partially functional.  This is app_mvc.py.  It uses a model/view/controller architecture and the corresponding files are in new directories controllers/models.  The old gui still works but I had to change some nomenclature to be consistent across the two.